requirements.txt
numpy
opencv-contrib-python
Pillow